Feat: Added User Condition Form.
In this commit, I introduced a new component that allows users to provide information about their physical condition. Currently, the structure of the form is in place. The form is intended to be shown only to new users during onboarding. Existing users with accounts will not need to complete the form.