Style: Revamped Color Palette and Navbar Styling.
In this commit, I introduced a new color palette to enhance contrast and visual appeal. The new colors complement each other well. Additionally, I applied a fresh style to the navbar, making navigation within the app smoother and more intuitive.

For the upcoming commit, I plan to implement a validation process for exiting the app when the logout button is touched.